requirements.txt
streamlit
easyocr
numpy
Pillow
torch
torchvision